[Chore #160905461] fixed errors after refactor
