+++ app/__init__.py
@@ -7,11 +7,11 @@ from flask_restful import Api
 # from instance.config import app_config
 from app.api.v1.views import Home, Orders, MyOrder
 from app.api.v1.auth import Reg, Login, Refresh
-from app.api.v2.auth import Signup, Loginv2
+from app.api.v2.views.auth import Signup, Loginv2
 # from app.api.v2.menu_view import Menu, TestMe, MenuItem
-from app.api.v2.view_menu import MenuView, MenuItem
-from app.api.v2.view_order import OrdersView, OrderItem
-from app.api.v2.view_user_order import UserOrder
+from app.api.v2.views.view_menu import MenuView, MenuItem
+from app.api.v2.views.view_order import OrdersView, OrderItem
+from app.api.v2.views.view_user_order import UserOrder
 
 
 api_bp = Blueprint('api', __name__)
