+++ app/api/v2/models/modelorder.py
@@ -1,6 +1,6 @@
 import os, sys
 LOCALPATH = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, LOCALPATH + '/../../../')
+sys.path.insert(0, LOCALPATH + '/../../../../')
 
 from app.api.v2.db import Db1
