+++ app/api/v2/views/auth.py
@@ -8,9 +8,9 @@ from flask_jwt_extended import (
 
 import os, sys
 LOCALPATH = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, LOCALPATH + '/../../../')
+sys.path.insert(0, LOCALPATH + '/../../../../')
 
-from app.api.v2.modeluser import UserModel
+from app.api.v2.models.modeluser import UserModel
