+++ app/api/v2/views/view_user_order.py
@@ -4,9 +4,9 @@ from flask_restful import Resource, reqparse
 from run import *
 import os, sys
 LOCALPATH = os.path.dirname(os.path.abspath(__file__))
-sys.path.insert(0, LOCALPATH + '/../../../')
+sys.path.insert(0, LOCALPATH + '/../../../../')
 
-from app.api.v2.modelorder import OrderModel
+from app.api.v2.models.modelorder import OrderModel
 
 parser = reqparse.RequestParser(bundle_errors=True)
 parser.add_argument('description', type=str, location='json')
